pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "closeloop"
version = "0.1.0"
description = "Deterministic multi-agent sales workflow scaffold"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "streamlit>=1.35.0",
  "google-generativeai>=0.8.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
